requirements.txt
streamlit
selenium
beautifulsoup4
requests
mysql-connector
pika
Flask
pandas